+++ src/twg2/text/tokenizer/CharParserFactory.java
@@ -4,7 +4,7 @@ import twg2.parser.condition.ParserFactory;
 import twg2.parser.condition.text.CharParser;
 import twg2.parser.textParser.TextParser;
 
-/** A CharParserFactory, commonly used by a CodeTokenizer or {@link CharMultiConditionParser}.<br>
+/** A CharParserFactory, commonly used by a {@link CharMultiConditionParser}.<br>
  * This interface provides an {@link #isMatch(char, TextParser)} method to check if a single character matches
  * the beginning of this CharParserFactory and a factory method to create a {@link CharParser}
  * designed so implementers can create recyclable, precondition filters, managed by a factory,
@@ -17,8 +17,9 @@ public interface CharParserFactory extends ParserFactory<CharParser> {
 
 	/** Check if a character matches the beginning of this {@code CharParserFactory}
 	 * @param ch the character
+	 * @param buf the {@link TextParser} source that {@code ch} was retrieved from
 	 * @return true if the character matches this precondition, false if not
 	 */
-	public boolean isMatch(char ch, TextParser pos);
+	public boolean isMatch(char ch, TextParser buf);
 
 }