+++ src/twg2/text/tokenizer/StringBoundedParserBuilder.java
@@ -4,12 +4,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map.Entry;
 
-import lombok.val;
 import twg2.collections.primitiveCollections.CharArrayList;
 import twg2.parser.Inclusion;
 import twg2.parser.condition.text.CharParser;
 import twg2.parser.condition.text.CharParserMatchable;
 import twg2.parser.condition.text.CharParserPredicate;
+import twg2.text.tokenizer.CharConditionPipe.AllRequired;
 import twg2.tuple.Tuples;
 
 /**
@@ -35,59 +35,59 @@ public class StringBoundedParserBuilder {
 
 
 	public StringBoundedParserBuilder addStartEndMarkers(String name, String start, String end, Inclusion includeEnd) {
-		val startFilter = new StringConditions.Start(name + "-start", new String[] { start }, Inclusion.INCLUDE);
-		val endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
-		val cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
+		StringConditions.Start startFilter = new StringConditions.Start(name + "-start", new String[] { start }, Inclusion.INCLUDE);
+		StringConditions.End endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
+		AllRequired<StringConditions.End> cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addStartEndMarkers(String name, String start, char end, Inclusion includeEnd) {
-		val startFilter = new StringConditions.Start(name + "-start", new String[] { start }, Inclusion.INCLUDE);
-		val endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
-		val cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
+		StringConditions.Start startFilter = new StringConditions.Start(name + "-start", new String[] { start }, Inclusion.INCLUDE);
+		CharConditions.End endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
+		AllRequired<CharConditions.End> cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addStartEndMarkers(String name, char start, String end, Inclusion includeEnd) {
-		val startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
-		val endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
-		val cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
+		CharConditions.Start startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
+		StringConditions.End endFilter = new StringConditions.End(name + "-end", new String[] { end }, includeEnd);
+		AllRequired<StringConditions.End> cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addStartEndMarkers(String name, char start, char end, Inclusion includeEnd) {
-		val startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
-		val endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
-		val cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
+		CharConditions.Start startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
+		CharConditions.End endFilter = new CharConditions.End(name + "-end", CharArrayList.of(end), includeEnd);
+		AllRequired<CharConditions.End> cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addStartEndNotPrecededByMarkers(String name, char start, char notPreced, char end, Inclusion includeEnd) {
-		val startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
-		val endFilter = new CharConditions.EndNotPrecededBy(name + "-end", CharArrayList.of(end), includeEnd, CharArrayList.of(notPreced));
-		val cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
+		CharConditions.Start startFilter = new CharConditions.Start(name + "-start", CharArrayList.of(start), Inclusion.INCLUDE);
+		CharConditions.EndNotPrecededBy endFilter = new CharConditions.EndNotPrecededBy(name + "-end", CharArrayList.of(end), includeEnd, CharArrayList.of(notPreced));
+		AllRequired<CharConditions.EndNotPrecededBy> cond = new CharConditionPipe.AllRequired<>(name, startFilter, endFilter);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addStringLiteralMarker(String name, String str) {
-		val cond = new StringConditions.Literal(name, new String[] { str }, Inclusion.INCLUDE);
+		StringConditions.Literal cond = new StringConditions.Literal(name, new String[] { str }, Inclusion.INCLUDE);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}
 
 
 	public StringBoundedParserBuilder addCharLiteralMarker(String name, char ch) {
-		val cond = new CharConditions.Literal(name, CharArrayList.of(ch), Inclusion.INCLUDE);
+		CharConditions.Literal cond = new CharConditions.Literal(name, CharArrayList.of(ch), Inclusion.INCLUDE);
 		this.filters.add(Tuples.of(cond.getFirstCharMatcher(), cond));
 		return this;
 	}