+++ src/twg2/text/tokenizer/StringConditions.java
@@ -3,7 +3,6 @@ package twg2.text.tokenizer;
 import java.util.Arrays;
 import java.util.Collection;
 
-import lombok.val;
 import twg2.arrays.ArrayUtil;
 import twg2.collections.dataStructures.Bag;
 import twg2.parser.Inclusion;
@@ -12,9 +11,13 @@ import twg2.parser.condition.text.CharParserMatchable;
 import twg2.parser.condition.text.CharParserPredicate;
 import twg2.parser.textFragment.TextFragmentRef;
 import twg2.parser.textFragment.TextFragmentRefImplMut;
+import twg2.parser.textParser.ParserPos;
 import twg2.parser.textParser.TextParser;
 
-/**
+/** {@link CharParserMatchable} string conditions, currently includes:<br>
+ *  - Literal<br>
+ *  - Start<br>
+ *  - End<br>
  * @author TeamworkGuy2
  * @since 2015-2-13
  */
@@ -33,8 +36,6 @@ public class StringConditions {
 		boolean failed = false;
 		/** count all accepted characters (including characters not explicitly part of 'matchingChars') */
 		int acceptedCount = 0;
-		/** count accepted characters (only from 'matchingChars') */
-		int matchCount = 0;
 		Inclusion includeMatchInRes;
 		StringBuilder dstBuf = new StringBuilder();
 		TextFragmentRefImplMut coords = new TextFragmentRefImplMut();
@@ -120,6 +121,12 @@ public class StringConditions {
 		}
 
 
+		@Override
+		public String toString() {
+			return "one " + Arrays.toString(this.originalStrs);
+		}
+
+
 		// package-private
 		void reset() {
 			matchingStrs.clearAndAddAll(originalStrs);
@@ -128,13 +135,36 @@ public class StringConditions {
 			dstBuf.setLength(0);
 			coords = new TextFragmentRefImplMut();
 			acceptedCount = 0;
-			matchCount = 0;
 		}
 
 
-		@Override
-		public String toString() {
-			return "one " + Arrays.toString(this.originalStrs);
+		/** Remove {@code matches} who's {@link String#charAt(int)} {@code off} do not match {@code ch}
+		 * @return 0 if no match, 1 if match found, 2 if match completed
+		 */
+		// package-private
+		static byte updateMatches(char ch, int off, Bag<String> matches) {
+			byte found = 0;
+			// reverse iterate through the bag so we don't have to adjust the loop counter when we remove elements
+			for(int i = matches.size() - 1; i > -1; i--) {
+				String str = matches.get(i);
+				int strLen = str.length();
+				// ignore string shorter than the current search offset (technically, if the precondition filter starts at offset 0, none of these should exist)
+				if(strLen > off) {
+					if(str.charAt(off) != ch) {
+						matches.remove(i);
+					}
+					else {
+						found = 1;
+						if(strLen == off + 1) {
+							found = 2;
+						}
+					}
+				}
+				else {
+					matches.remove(i);
+				}
+			}
+			return found;
 		}
 
 
@@ -149,7 +179,7 @@ public class StringConditions {
 
 
 
-	/**
+	/** A matcher that finds exact string sequences
 	 */
 	public static class Literal extends Start {
 
@@ -160,8 +190,7 @@ public class StringConditions {
 
 		@Override
 		public Literal copy() {
-			val copy = new Literal(name, originalStrs, includeMatchInRes);
-			return copy;
+			return new Literal(name, originalStrs, includeMatchInRes);
 		}
 
 	}
@@ -169,7 +198,7 @@ public class StringConditions {
 
 
 
-	/**
+	/** A matcher that finds starts-with sequences
 	 */
 	public static class Start extends BaseStringParser {
 
@@ -185,35 +214,16 @@ public class StringConditions {
 				super.failed = true;
 				return false;
 			}
-			boolean anyFound = false;
-			Bag<String> matchingStrs = super.matchingStrs;
-			// reverse iterate through the bag so we don't have to adjust the loop counter when we remove elements
-			for(int i = matchingStrs.size() - 1; i > -1; i--) {
-				String strI = matchingStrs.get(i);
-				int strILen = strI.length();
-				// ignore string shorter than the current search offset (technically, if the precondition filter starts at offset 0, none of these should exist)
-				if(strILen > off) {
-					if(strI.charAt(off) != ch) {
-						matchingStrs.remove(i);
-					}
-					else {
-						anyFound = true;
-						if(strILen == off + 1) {
-							super.anyComplete = true;
-						}
-					}
-				}
-				else {
-					matchingStrs.remove(i);
-				}
+			byte found = updateMatches(ch, off, super.matchingStrs);
+			if(found == 2) {
+				super.anyComplete = true;
 			}
 
-			if(anyFound) {
+			if(found > 0) {
 				if(super.acceptedCount == 0) {
 					super.coords.setStart(buf);
 				}
 				super.acceptedCount++;
-				super.matchCount++;
 				super.dstBuf.append(ch);
 				if(super.anyComplete) {
 					super.coords.setEnd(buf);
@@ -230,8 +240,7 @@ public class StringConditions {
 
 		@Override
 		public Start copy() {
-			val copy = new Start(super.name, super.originalStrs, super.includeMatchInRes);
-			return copy;
+			return new Start(super.name, super.originalStrs, super.includeMatchInRes);
 		}
 
 	}
@@ -239,7 +248,7 @@ public class StringConditions {
 
 
 
-	/**
+	/** A matcher that finds ends-with sequences
 	 */
 	public static class End extends BaseStringParser {
 
@@ -252,35 +261,22 @@ public class StringConditions {
 		public boolean acceptNext(char ch, TextParser buf) {
 			int off = super.dstBuf.length();
 			if(super.isComplete()) {
-				super.failed = true;
-				return false;
-			}
-			boolean anyFound = false;
-			Bag<String> matchingStrs = super.matchingStrs;
-			// reverse iterate through the bag so we don't have to adjust the loop counter when we remove elements
-			for(int i = matchingStrs.size() - 1; i > -1; i--) {
-				String strI = matchingStrs.get(i);
-				int strILen = strI.length();
-				// ignore string shorter than the current search offset (technically, if the precondition filter starts at offset 0, none of these should exist)
-				if(strILen > off) {
-					if(strI.charAt(off) != ch) {
-						matchingStrs.remove(i);
-					}
-					else {
-						anyFound = true;
-						if(strILen == off + 1) {
-							super.anyComplete = true;
-						}
-					}
+				// if this end sequence is still matched by adding the next character
+				if(findMoreRecentMatch(ch, buf)) {
+					return true;
 				}
 				else {
-					matchingStrs.remove(i);
+					super.failed = true;
+					return false;
 				}
 			}
 
-			super.matchCount++;
+			byte found = updateMatches(ch, off, super.matchingStrs);
+			if(found == 2) {
+				super.anyComplete = true;
+			}
 
-			if(anyFound) {
+			if(found > 0) {
 				if(super.acceptedCount == 0) {
 					super.coords.setStart(buf);
 				}
@@ -291,7 +287,10 @@ public class StringConditions {
 				}
 			}
 			else {
-				super.reset();
+				boolean match = findMoreRecentMatch(ch, buf);
+				if(!match) {
+					super.reset();
+				}
 			}
 			return true;
 		}
@@ -299,8 +298,55 @@ public class StringConditions {
 
 		@Override
 		public End copy() {
-			val copy = new End(super.name, super.originalStrs, super.includeMatchInRes);
-			return copy;
+			return new End(super.name, super.originalStrs, super.includeMatchInRes);
+		}
+
+
+		/** Check for a shorter matching sequence in currently matched chars and update this end condition's buffer and coords to start at the shorter sub-match
+		 * @return true if a shorter match exists
+		 */
+		private boolean findMoreRecentMatch(char ch, TextParser buf) {
+			// start i = 1 because this method should only get called when current sequence is a match and another characters is available
+			for(int i = 1, size = super.dstBuf.length(); i < size; i++) {
+				this.partialReset();
+				for(int j = i; j < size; j++) {
+					byte found = updateMatches(super.dstBuf.charAt(j), j - i, super.matchingStrs);
+					if(found == 2) {
+						super.anyComplete = true;
+					}
+				}
+				// newest char
+				byte found = updateMatches(ch, size - i, super.matchingStrs);
+				if(found == 2) {
+					super.anyComplete = true;
+				}
+				// found match
+				if(super.matchingStrs.size() > 0) {
+					super.dstBuf.delete(0, i);
+					super.dstBuf.append(ch);
+					super.acceptedCount = size - i + 1;
+					int pos = super.coords.getOffsetStart() + i;
+					int lineNum = buf.getLineNumbers().getLineNumber(pos);
+					int columnNum = pos - buf.getLineNumbers().getLineOffset(lineNum);
+					super.coords.setStart(new ParserPos.Impl(pos, lineNum + 1, columnNum + 1));
+					if(super.anyComplete) {
+						super.coords.setEnd(buf);
+					}
+					return true;
+				}
+			}
+			return false;
+		}
+
+
+		// package-private
+		void partialReset() {
+			matchingStrs.clearAndAddAll(originalStrs);
+			anyComplete = false;
+			failed = false;
+			//dstBuf.setLength(0);
+			//coords = new TextFragmentRefImplMut();
+			acceptedCount = 0;
 		}
 
 	}