+++ test/twg2/text/tokenizer/test/CharConditionPipeFilterTest.java
@@ -7,8 +7,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import lombok.val;
-
 import org.junit.Test;
 
 import twg2.collections.primitiveCollections.CharArrayList;
@@ -26,21 +24,21 @@ public class CharConditionPipeFilterTest {
 
 	@Test
 	public void allrequiredTest() throws IOException {
-		val name = "AllRequired";
-		val condSet0 = new ArrayList<CharParser>(Arrays.asList(
+		String name = "AllRequired";
+		List<CharParser> condSet0 = new ArrayList<>(Arrays.asList(
 				new CharConditions.Start("<", CharArrayList.of('<'), Inclusion.INCLUDE),
-				new CharConditions.End("<", CharArrayList.of('>'), Inclusion.INCLUDE)
+				new CharConditions.End(">", CharArrayList.of('>'), Inclusion.INCLUDE)
 		));
 
-		val condSet1 = new ArrayList<CharParser>(Arrays.asList(
+		List<CharParser> condSet1 = new ArrayList<>(Arrays.asList(
 				new StringConditions.Literal("test", new String[] { "test" }, Inclusion.INCLUDE)
 		));
 
-		val condSet2 = new ArrayList<CharParser>(Arrays.asList(
+		List<CharParser> condSet2 = new ArrayList<>(Arrays.asList(
 				new StringConditions.End("!", new String[] { "!" }, Inclusion.INCLUDE)
 		));
 
-		val condSets = new ArrayList<List<CharParser>>(Arrays.asList(
+		List<List<CharParser>> condSets = new ArrayList<>(Arrays.asList(
 				condSet0,
 				condSet1,
 				condSet2
@@ -63,22 +61,22 @@ public class CharConditionPipeFilterTest {
 
 	@Test
 	public void optionalSuffixTest() throws IOException {
-		val name = "OptionalSuffix";
-		val condSet0 = new ArrayList<CharParser>(Arrays.asList(
+		String name = "OptionalSuffix";
+		List<CharParser> condSet0 = new ArrayList<>(Arrays.asList(
 				new CharConditions.Start("<", CharArrayList.of('<'), Inclusion.INCLUDE),
 				new CharConditions.End(">", CharArrayList.of('>'), Inclusion.INCLUDE)
 		));
 
-		val condSetOptional = new ArrayList<CharParser>(Arrays.asList(
+		List<CharParser> condSetOptional = new ArrayList<>(Arrays.asList(
 				new StringConditions.Literal("test", new String[] { "test" }, Inclusion.INCLUDE)
 		));
 
-		val condSets = new ArrayList<List<CharParser>>(Arrays.asList(
+		List<List<CharParser>> condSets = new ArrayList<>(Arrays.asList(
 				condSet0,
 				condSetOptional
 		));
 
-		val pipeCond = new CharConditionPipe.OptionalSuffix<CharParser>(name, condSets);
+		CharConditionPipe.OptionalSuffix<CharParser> pipeCond = new CharConditionPipe.OptionalSuffix<CharParser>(name, condSets);
 		// TODO testing CharConditionPipe.setupPipeOptionalSuffixFilter(pipeCond);
 
 		parseTest(true, false, name, pipeCond, "<abc>");
@@ -93,22 +91,22 @@ public class CharConditionPipeFilterTest {
 
 	@Test
 	public void repeatableSeparatedText() {
-		val name = "RepeatableSeparator";
-		val condSet0 = new ArrayList<CharParser>(Arrays.asList(
+		String name = "RepeatableSeparator";
+		List<CharParser> condSet0 = new ArrayList<CharParser>(Arrays.asList(
 				new CharConditions.Start("<", CharArrayList.of('<'), Inclusion.INCLUDE),
 				new CharConditions.End(">", CharArrayList.of('>'), Inclusion.INCLUDE)
 		));
 
-		val condSetOptional = new ArrayList<CharParser>(Arrays.asList(
+		List<CharParser> condSetOptional = new ArrayList<CharParser>(Arrays.asList(
 				new StringConditions.Literal("separator", new String[] { ", " }, Inclusion.INCLUDE)
 		));
 
-		val condSets = new ArrayList<List<CharParser>>(Arrays.asList(
+		List<List<CharParser>> condSets = new ArrayList<>(Arrays.asList(
 				condSet0,
 				condSetOptional
 		));
 
-		val pipeCond = new CharConditionPipe.RepeatableSeparator<CharParser>(name, condSets);
+		CharConditionPipe.RepeatableSeparator<CharParser> pipeCond = new CharConditionPipe.RepeatableSeparator<CharParser>(name, condSets);
 		// TODO testing CharConditionPipe.setupPipeRepeatableSeparatorFilter(pipeCond);
 
 		parseTest(true, false, name, pipeCond, "<abc>");